[Refactoring] #91 - Mapper 로 Conversion로직 분리 및 코드 정리
